removing autoscale annotation and related code
